Flex 06 accomplished
tbh i don't think i'm gonna do a true end for this
one, it looks darn good. Only issue really is the
empty space infront the first links in the header
and footer. Besides that:

Removed the links, flattened the list
with flex and made the links on top and bottom by
making the body a flex collum.

centered the stuff in the centered

space between for the links
note: containr, not ul itself