Flex 7 true ending
yeah so i worked on it and stuff and it looks
good so i'm calling it here. of note:

1. Only real things for wrap is a contianer
with flex wrap on and the items inside to have
a set width.

2. to do the footer better, make the middle
section a flex 1 1 auto. This lets the middle
expand to fill the available space, which ipso
facto, pins the footer to the bottom.

besides that just decoration stuff cool cool.